IGNITE-22732 Code review changes
