+++ modules/calcite/src/test/java/org/apache/ignite/internal/processors/query/calcite/integration/DataTypesTest.java
@@ -29,8 +29,6 @@ import org.apache.calcite.tools.Frameworks;
 import org.apache.ignite.IgniteCache;
 import org.apache.ignite.cache.QueryEntity;
 import org.apache.ignite.internal.processors.query.IgniteSQLException;
-import org.apache.ignite.internal.processors.query.calcite.AbstractTransactionalSqlTest.SqlTransactionMode;
-import org.apache.ignite.internal.processors.query.calcite.AbstractTransactionalSqlTest.SupplierX;
 import org.apache.ignite.internal.util.typedef.F;
 import org.junit.Test;
